fix: separate galactica
