+++ src/Components/Reusable/GasFeeSpeed/GasFeeSpeed.tsx
@@ -1,20 +1,16 @@
-import { ethers } from "ethers"
 import moment from "moment"
-import React, { ComponentClass, FC, useCallback, useEffect, useMemo, useState } from "react"
+import React, { useCallback, useEffect, useMemo, useState } from "react"
 import { StyleSheet } from "react-native"
-import Animated from "react-native-reanimated"
 import { useInterval } from "usehooks-ts"
-import { useExchangeRate } from "~Api/Coingecko"
 import { BaseButton, BaseCard, BaseIcon, BaseSpacer, BaseText, BaseView } from "~Components/Base"
-import { ColorThemeType, GasPriceCoefficient, getCoinGeckoIdBySymbol, VTHO } from "~Constants"
-import { useBlinkStyles, useBottomSheetModal, useFormatFiat, useThemedStyles } from "~Hooks"
+import { ColorThemeType, GasPriceCoefficient } from "~Constants"
+import { useBottomSheetModal, useThemedStyles } from "~Hooks"
 import { TransactionFeesResult } from "~Hooks/useTransactionFees/useTransactionFees"
 import { useI18nContext } from "~i18n"
-import { selectCurrency, useAppSelector } from "~Storage/Redux"
-import { BigNutils } from "~Utils"
-import { TokenImage } from "../TokenImage"
 import { SPEED_MAP } from "./constants"
+import { GalacticaEstimation } from "./GalacticaEstimation"
 import { GasFeeSpeedBottomSheet } from "./GasFeeSpeedBottomSheet"
+import { LegacyEstimation } from "./LegacyEstimation"
 
 type Props = {
     options: TransactionFeesResult
@@ -22,20 +18,8 @@ type Props = {
     selectedFeeOption: GasPriceCoefficient
     onRefreshFee: () => void
     gasUpdatedAt: number
+    isGalactica?: boolean
 }
-
-const wrapFunctionComponent = <TProps,>(Component: FC<TProps>): ComponentClass<TProps> =>
-    class extends React.Component<TProps> {
-        constructor(props: TProps) {
-            super(props)
-        }
-
-        render() {
-            return <Component {...this.props} />
-        }
-    }
-
-const AnimatedText = Animated.createAnimatedComponent(wrapFunctionComponent(BaseText))
 
 export const GasFeeSpeed = ({
     options,
@@ -43,39 +27,16 @@ export const GasFeeSpeed = ({
     selectedFeeOption,
     onRefreshFee,
     gasUpdatedAt,
+    isGalactica,
 }: Props) => {
     const { LL } = useI18nContext()
     const { theme, styles } = useThemedStyles(baseStyles)
 
-    const { formatValue, formatFiat } = useFormatFiat()
-
     const { onClose, onOpen, ref } = useBottomSheetModal()
-
-    const currency = useAppSelector(selectCurrency)
 
     const [secondsRemaining, setSecondsRemaining] = useState(10)
 
-    const { data: exchangeRate } = useExchangeRate({
-        id: getCoinGeckoIdBySymbol[VTHO.symbol],
-        vs_currency: currency,
-    })
-
     const { estimatedFee, maxFee } = useMemo(() => options[selectedFeeOption], [options, selectedFeeOption])
-
-    const estimatedFeeVtho = useMemo(
-        () => parseFloat(ethers.utils.formatEther(estimatedFee.toString)),
-        [estimatedFee.toString],
-    )
-    const maxFeeVtho = useMemo(() => parseFloat(ethers.utils.formatEther(maxFee.toString)), [maxFee.toString])
-
-    const estimatedFeeFiat = useMemo(() => {
-        return BigNutils().toCurrencyConversion(estimatedFeeVtho.toString() || "0", exchangeRate ?? 1)
-    }, [exchangeRate, estimatedFeeVtho])
-
-    const estimatedFormattedFiat = useMemo(() => {
-        if (estimatedFeeFiat.isLeesThan_0_01) return formatFiat({ amount: 0.01 })
-        return formatFiat({ amount: parseInt(estimatedFeeFiat.preciseValue, 10) })
-    }, [estimatedFeeFiat.isLeesThan_0_01, estimatedFeeFiat.preciseValue, formatFiat])
 
     const intervalFn = useCallback(() => {
         setSecondsRemaining(Math.floor(moment(gasUpdatedAt).add(10, "seconds").diff(moment(), "seconds")))
@@ -86,8 +47,6 @@ export const GasFeeSpeed = ({
     useEffect(() => {
         if (secondsRemaining === 0) onRefreshFee()
     }, [onRefreshFee, secondsRemaining])
-
-    const blinkStyles = useBlinkStyles({ enabled: secondsRemaining <= 3, duration: 1000 })
 
     return (
         <BaseView flexDirection="column" gap={16} mt={16}>
@@ -132,47 +91,19 @@ export const GasFeeSpeed = ({
                     </BaseButton>
                 </BaseView>
                 <BaseSpacer height={1} background={theme.colors.pressableCardBorder} />
-                <BaseView flexDirection="column" style={styles.section} gap={4}>
-                    <BaseView flexDirection="row" justifyContent="space-between" w={100}>
-                        <BaseText color={theme.colors.textLight} typographyFont="captionMedium">
-                            {LL.ESTIMATED_FEE()}
-                        </BaseText>
-                        <BaseText color={theme.colors.textLight} typographyFont="captionMedium">
-                            {LL.MAX_FEE()}
-                        </BaseText>
-                    </BaseView>
-                    <BaseView flexDirection="row" justifyContent="space-between" w={100} alignItems="center">
-                        <BaseView flexDirection="row" gap={8}>
-                            <TokenImage icon={VTHO.icon} isVechainToken iconSize={16} />
-                            <AnimatedText
-                                typographyFont="subSubTitleBold"
-                                color={theme.colors.assetDetailsCard.title}
-                                style={blinkStyles}>
-                                {VTHO.symbol}
-                            </AnimatedText>
-                            <AnimatedText
-                                typographyFont="subSubTitleBold"
-                                color={theme.colors.assetDetailsCard.title}
-                                style={blinkStyles}>
-                                {formatValue(estimatedFeeVtho)}
-                            </AnimatedText>
-                            <AnimatedText
-                                typographyFont="bodyMedium"
-                                color={theme.colors.textLight}
-                                style={blinkStyles}>
-                                {estimatedFeeFiat.isLeesThan_0_01
-                                    ? `< ${estimatedFormattedFiat}`
-                                    : estimatedFormattedFiat}
-                            </AnimatedText>
-                        </BaseView>
-                        <AnimatedText
-                            typographyFont="subSubTitleBold"
-                            color={theme.colors.textLight}
-                            style={blinkStyles}>
-                            {formatValue(maxFeeVtho)} {VTHO.symbol}
-                        </AnimatedText>
-                    </BaseView>
-                </BaseView>
+                {isGalactica ? (
+                    <GalacticaEstimation
+                        options={options}
+                        selectedFeeOption={selectedFeeOption}
+                        secondsRemaining={secondsRemaining}
+                    />
+                ) : (
+                    <LegacyEstimation
+                        options={options}
+                        selectedFeeOption={selectedFeeOption}
+                        secondsRemaining={secondsRemaining}
+                    />
+                )}
                 <GasFeeSpeedBottomSheet
                     ref={ref}
                     estimatedFee={estimatedFee}
